+++ src/common/dsp/SampleAndHoldOscillator.cpp
@@ -197,7 +197,7 @@ void SampleAndHoldOscillator::convolute(int voice, bool FM, bool stereo)
    float wf = l_shape.v * 0.8 * invertcorrelation;
    float wfabs = fabs(wf);
    float smooth = l_smooth.v;
-   float rand11 = (((float)rand() * rcp(RAND_MAX)) * 2.f - 1.f);
+   float rand11 = (((float)rand() * rcp((float)RAND_MAX)) * 2.f - 1.f);
    float randt = rand11 * (1 - wfabs) - wf * last_level[voice];
 
    randt = randt * rcp(1.0f - wfabs);
